+++ app/Http/Controllers/Back/InvoicePaymentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Back;
 use App\CashAccount;
 use App\Invoice;
 use App\InvoicePayment;
+use App\UserRole;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -45,6 +46,9 @@ class InvoicePaymentController extends Controller
      */
     public function index(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $start = ($request->input('checkin_date')) ? $request->input('checkin_date') : date('Y-m-d', strtotime("-1 month"));
         $end = ($request->input('checkout_date')) ? $request->input('checkout_date') : date('Y-m-d');
         $status = ($request->input('status')) ? $request->input('status') : 0;
@@ -64,6 +68,9 @@ class InvoicePaymentController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['cash_account'] = CashAccount::all();
         $data['invoice'] = Invoice::where('paid', 0)->get();
         $data['parent_menu'] = $this->parent;
@@ -78,6 +85,9 @@ class InvoicePaymentController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'payment_date'  => 'required',
             'invoice_id'  => 'required',
@@ -122,6 +132,9 @@ class InvoicePaymentController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['invoice'] = Invoice::where('paid', 0)->get();
         $data['parent_menu'] = $this->parent;
         $data['row'] = $this->model->find($id);
@@ -137,6 +150,9 @@ class InvoicePaymentController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'payment_date'  => 'required',
             'invoice_id'  => 'required',
@@ -174,6 +190,9 @@ class InvoicePaymentController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function changeStatus($id, $status) {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         if($status == 1){
@@ -201,6 +220,9 @@ class InvoicePaymentController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 15, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
