+++ app/Http/Controllers/Back/MutationController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Back;
 
 use App\CashAccount;
 use App\Mutation;
+use App\UserRole;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -44,6 +45,9 @@ class MutationController extends Controller
      */
     public function index(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'read')){
+            return view("auth.unauthorized");
+        }
         $start = ($request->input('checkin_date')) ? $request->input('checkin_date') : date('Y-m-d', strtotime("-1 month"));
         $end = ($request->input('checkout_date')) ? $request->input('checkout_date') : date('Y-m-d');
         $status = ($request->input('status')) ? $request->input('status') : 0;
@@ -64,6 +68,9 @@ class MutationController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['cash_account'] = CashAccount::all();
         $data['parent_menu'] = $this->parent;
         return view("back.".$this->module.".create", $data);
@@ -77,6 +84,9 @@ class MutationController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'amount'  => 'required',
@@ -117,6 +127,9 @@ class MutationController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['cash_account'] = CashAccount::all();
         $data['row'] = $this->model->find($id);
@@ -132,6 +145,9 @@ class MutationController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'amount'  => 'required',
@@ -169,6 +185,9 @@ class MutationController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function changeStatus($id, $status) {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         if($status == 1){
@@ -190,6 +209,9 @@ class MutationController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 16, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
