+++ app/Http/Controllers/HouseKeep/FoundController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\HouseKeep;
 
 use App\Found;
 use App\Lost;
+use App\UserRole;
 use Illuminate\Foundation\Auth\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -51,6 +52,9 @@ class FoundController extends Controller
      */
     public function index(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'read')){
+            return view("auth.unauthorized");
+        }
         $founder_name = $request->input('founder_name');
         $item_name = $request->input('item_name');
         $rows = $this->model->where('founder_name', 'LIKE', "%$founder_name%")->where('item_name', 'LIKE', "%$item_name%")->paginate();
@@ -68,6 +72,9 @@ class FoundController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['lost'] = Lost::all();
         $data['EmployeeModel'] = new User();
@@ -83,6 +90,9 @@ class FoundController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'founder_name' => 'required'
@@ -124,6 +134,9 @@ class FoundController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['lost'] = Lost::all();
         $data['EmployeeModel'] = new User();
@@ -141,6 +154,9 @@ class FoundController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'founder_name' => 'required'
@@ -169,6 +185,9 @@ class FoundController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function changeStatus($id, $status) {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         $data->status = $status;
@@ -184,6 +203,9 @@ class FoundController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
