+++ app/Http/Controllers/HouseKeep/LostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\HouseKeep;
 
 use App\Lost;
 use App\Guest;
+use App\UserRole;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -51,6 +52,9 @@ class LostController extends Controller
      */
     public function index(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'read')){
+            return view("auth.unauthorized");
+        }
         $report_name = $request->input('report_name');
         $item_name = $request->input('item_name');
         $rows = $this->model->where('report_name', 'LIKE', "%$report_name%")->where('item_name', 'LIKE', "%$item_name%")->paginate();
@@ -68,6 +72,9 @@ class LostController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['guestModel'] = new Guest();
         $data['guest'] = $this->guest;
@@ -82,6 +89,9 @@ class LostController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'report_name' => 'required'
@@ -125,6 +135,9 @@ class LostController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['guestModel'] = new Guest();
         $data['guest'] = $this->guest;
@@ -141,6 +154,9 @@ class LostController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'date'  => 'required',
             'report_name' => 'required'
@@ -171,6 +187,9 @@ class LostController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function changeStatus($id, $status) {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         $data->status = $status;
@@ -186,6 +205,9 @@ class LostController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 11, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
