+++ app/Http/Controllers/Master/BankController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Master;
 
 use App\Bank;
 use App\Menu;
+use App\UserRole;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Helpers\GlobalHelper;
@@ -43,6 +44,9 @@ class BankController extends Controller
      */
     public function index()
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'read')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $rows = $this->model->paginate();
         $data['rows'] = $rows;
@@ -56,6 +60,9 @@ class BankController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         return view("master.".$this->module.".create", $data);
     }
@@ -68,6 +75,9 @@ class BankController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'bank_name'  => 'required|max:75|min:3'
         ]);
@@ -99,6 +109,9 @@ class BankController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['row'] = $this->model->find($id);
         return view("master.".$this->module.".edit", $data);
@@ -113,6 +126,10 @@ class BankController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
+
         $this->validate($request,[
             'bank_name'  => 'required|max:75|min:3'
         ]);
@@ -144,6 +161,9 @@ class BankController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function changeStatus($id, $status) {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         if($status == 1){
@@ -165,6 +185,9 @@ class BankController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
