+++ app/Http/Controllers/Master/CashAccountController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Master;
 
 use App\CashAccount;
+use App\UserRole;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Helpers\GlobalHelper;
@@ -42,6 +43,9 @@ class CashAccountController extends Controller
      */
     public function index()
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'read')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $rows = $this->model->paginate();
         $data['rows'] = $rows;
@@ -55,6 +59,9 @@ class CashAccountController extends Controller
      */
     public function create()
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         return view("master.".$this->module.".create", $data);
     }
@@ -67,6 +74,9 @@ class CashAccountController extends Controller
      */
     public function store(Request $request)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'create')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'cash_account_name'  => 'required|max:75|min:3',
             'cash_account_desc'  => 'max:255|min:3',
@@ -102,6 +112,9 @@ class CashAccountController extends Controller
      */
     public function edit($id)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data['parent_menu'] = $this->parent;
         $data['row'] = $this->model->find($id);
         return view("master.".$this->module.".edit", $data);
@@ -116,6 +129,9 @@ class CashAccountController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $this->validate($request,[
             'cash_account_name'  => 'required|max:75|min:3',
             'cash_account_desc'  => 'max:255|min:3',
@@ -152,6 +168,9 @@ class CashAccountController extends Controller
      */
     public function changeStatus($id, $status)
     {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'update')){
+            return view("auth.unauthorized");
+        }
         $data = $this->model->find($id);
 
         if ($status == 1) {
@@ -173,6 +192,9 @@ class CashAccountController extends Controller
      * @return \Illuminate\Http\RedirectResponse
      */
     public function softDelete($id) {
+        if(!UserRole::checkAccess($subModule = 3, $type = 'delete')){
+            return view("auth.unauthorized");
+        }
         $this->model->find($id)->delete();
         $message = GlobalHelper::setDisplayMessage('success', 'Success to delete data');
         return redirect(route($this->module.".index"))->with('displayMessage', $message);
